+++ fieldOrder.h
@@ -1,9 +1,9 @@
 // The fieldOrder struct contains ints for x, y, z in nanoTeslas,
 // as well as the int t in seconds. 
 typedef struct fieldOrderNode {
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
     int t;
     struct fieldOrderNode* next;
 } fieldOrderNode_t;
@@ -15,7 +15,7 @@ typedef struct fieldOrderQueue {
 
 fieldOrderQueue_t* fieldOrderQueue_init();
 
-fieldOrderNode_t* fieldOrderNode_create(int x, int y, int z, int t);
+fieldOrderNode_t* fieldOrderNode_create(float x, float y, float z, int t);
 void fieldOrderQueue_enqueue(fieldOrderQueue_t* queue, fieldOrderNode_t* node);
 fieldOrderNode_t* fieldOrderQueue_dequeue(fieldOrderQueue_t* queue);
